Added Code for the Q111 - Safety Checks
Q111 is for Manual Shutdown.  Right now it does nothing more than
checking the UART (B0, B1) for a response for response from the USER
(also tied to Q112 UART, but Q112 UART will never be in receive mode).
This will output to a GPIO (A0) which will be checked on the Q112 every
loop to continue running or stop.